refactor: store split to 3 different files
