+++ src/components/AddGroupModal.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent, FormEvent, useRef } from 'react';
 import { Button, Form, Modal } from 'react-bootstrap';
-import { groupsStore } from 'src/mobx/store';
+import { groupsStore } from 'src/mobx/groupsStore';
 
 interface IAddContactModal {
   handleClose: () => void;
