+++ src/pages/ContactPage.tsx
@@ -2,7 +2,7 @@ import { Col, Row } from 'react-bootstrap';
 import { useParams } from 'react-router-dom';
 import { ContactCard } from 'src/components/ContactCard';
 import { Empty } from 'src/components/Empty';
-import { contactsStore } from 'src/mobx/store';
+import { contactsStore } from 'src/mobx/contactsStore';
 
 export const ContactPage = () => {
   const { contactId } = useParams<{ contactId: string }>();
